BotAssistant/requirements.txt
py-cord
mysql-connector-python==8.0.29